Backend/requirements.txt
asgiref==3.11.0
async-generator==1.10
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==6.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
MarkupSafe==3.0.3
mypy_extensions==1.1.0
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
selenium==4.40.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0

gunicorn
whitenoise
dj-database-url
